+++ iidm/iidm-extensions/src/test/java/com/powsybl/iidm/network/extensions/MeasurementTest.java
@@ -1,3 +1,11 @@
+/**
+ * Copyright (c) 2025, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
 package com.powsybl.iidm.network.extensions;
 
 import java.util.Set;
@@ -11,6 +19,9 @@ import org.junit.jupiter.api.Test;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.ThreeSides;
 
+/**
+ * @author Laurent Garnier {@literal <laurent.garnier_externe at rte-france.com>}
+ */
 class MeasurementTest {
 
     @Test
