+++ iidm/iidm-tck/src/test/java/com/powsybl/iidm/network/tck/extensions/AbstractMeasurementsTest.java
@@ -7,6 +7,7 @@
  */
 package com.powsybl.iidm.network.tck.extensions;
 
+import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.Load;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.TwoWindingsTransformer;
@@ -115,5 +116,30 @@ public abstract class AbstractMeasurementsTest {
 
         twtMeasurements.cleanIfEmpty();
         assertNotNull(twt.getExtension(Measurements.class));
+    }
+
+    @Test
+    void setValueAndValidity() {
+        Network network = EurostagTutorialExample1Factory.create();
+        network.setCaseDate(ZonedDateTime.parse("2016-06-27T12:27:58.535+02:00"));
+
+        Load load = network.getLoad("LOAD");
+        load.newExtension(MeasurementsAdder.class).add();
+        Measurement measurement = load.getExtension(Measurements.class)
+                .newMeasurement()
+                .setType(Measurement.Type.OTHER)
+                .setValue(Double.NaN)
+                .setStandardDeviation(1.0)
+                .setValid(false)
+                .add();
+
+        measurement.setValueAndValidity(200.0d, true);
+        assertEquals(200.0d, measurement.getValue());
+        assertTrue(measurement.isValid());
+
+        measurement.setValueAndValidity(Double.NaN, false);
+        assertFalse(measurement.isValid());
+
+        assertThrows(PowsyblException.class, () -> measurement.setValueAndValidity(Double.NaN, true));
     }
 }
